Default assignments on classroom join
For "custom" (Wildcam-style, custom assignment) programs, teachers create assignments before or after a classroom is joined and then assign them to students. However, for non-custom (i2a) programs, the order of operations will look more like this:

1) Teacher creates classroom
2) Default assignments are automatically generated by the i2a frontend.
3) Stutent attempts to join classroom and is automatically assigned each of the classroom's assignments.

This PR adds that check, based on the program's `custom` attribute, and assigns them if it is false.