Refactor of assignment creation (#75)
Assignments weren't being created correctly on custom programs. Now, the assignment creation order of operations looks like this:

is the program custom (wildcam-style)?
    clone the workflow indicated in request
    get the new workflow's info from panoptes
    create subject set on that project
    link that subject set to the cloned workflow
    create assignment (incl. workflow_id = cloned workflow id and subject_set_id)
not custom? (i2a-alike)
   get workflow data from panoptes
   create assignment (incl. assignment.workflow_id = workflow id)

Required an additional call to panoptes because the project id is needed before the workflow is cloned, but not really an issue. Also required addition of a client method on the Panoptes client to add a subject set to a workflow, so the version is bumped.