Fix classroom creation
Attributes were being ignored because the classroom wasn't being saved after updating the object in the operation. This wasn't being caught by the spec because the test object wasn't being reloaded from the db before testing it. Also fixed the update op so that it won't update non-included attributes to nil.