Count classifications by single user group
Non-custom (i2a-style) projects include a "user group" key in the classification metadata that indicates the specific UG/classroom that should be incremented. This PR ensures that in these cases, only the single classroom's counter is incremented rather than the corresponding classroom for every non-ID user group the user is in. This latter case seems to be the desired behavior for Wildcam-style projects, so if there is no user group included in the classfication metadata (i.e. there was no query param on the URL during classification) then they'll all be ticked as before.